target published amo-tools-suite, Use js file for download-stats
