+++ package.json
@@ -26,7 +26,7 @@
     "postinstall": "electron-builder install-app-deps",
     "pack": "electron-builder --dir",
     "dist": "electron-builder",
-    "download-stats": "ts-node scripts/download-stats.ts"
+    "download-stats": "node scripts/download-stats.js"
   },
   "private": true,
   "dependencies": {
@@ -45,7 +45,7 @@
     "@ng-bootstrap/ng-bootstrap": "14.0.1",
     "@popperjs/core": "^2.11.2",
     "ajv": "8.12.0",
-    "amo-tools-suite": "../AMO-Tools-Suite",
+    "amo-tools-suite": "1.0.3",
     "billboard.js": "^3.7.2",
     "bootstrap": "4.3.1",
     "core-js": "3.21.1",
